Move restart spec to the config.types module.
